Merged revisions 1336,1338 via svnmerge from https://origsvn.digium.com/svn/asterisk-gui/branches/1.4
........
r1336 | bkruse | 2007-08-14 18:08:57 -0500 (Tue, 14 Aug 2007) | 1 line

The beginning of digital card support. [1] Added ability to check if any spans are on the computer via /etc/asterisk/ztscan.conf (generated by ztscan.conf). [2] With this, it checks the general section of the file, and if continue=no, the gui will tell the user no spans were found, and give the option to redirect them to the digital.html page (under constructions) which will guide them through a setup of their spans (if they have any) [3] This is the first of a couple of commits having to deal with digital card detection, but it is looking very promising.
........
r1338 | bkruse | 2007-08-14 18:10:28 -0500 (Tue, 14 Aug 2007) | 1 line

digital detection page placeholder, the place that trunks.html redirects if the user requests, after being told ztscan could not find cards/spans
........


git-svn-id: http://svn.digium.com/svn/asterisk-gui/trunk@1339 f2be6d49-a0cf-43a2-a903-cc364acb77d4